Fix database not getting reset between all tests
Fixtures are specific to the class, so instead split tests
into separate classes in order to clean up and rebuild the db between
each one.

I'm not sure if [Collection("Database")] can be moved to
DatabaseTestPreset so it doesn't need to be added to each test class.
This should be looked into.